% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-accessors.R
\name{baseCalled}
\alias{baseCalled}
\title{Extract baseCalled slot}
\usage{
baseCalled(x)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{Fast5Summary}}}
}
\value{
A data.table with 6 columns
}
\description{
This generic function accesses the baseCalled slot stored in an object 
derived from the Fast5Summary class.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   baseCalled( s.typhi.rep2 )
}
}

