% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-accessors.R
\name{eventData}
\alias{eventData}
\title{Extract eventData slot}
\usage{
eventData(x)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{Fast5Summary}}}
}
\value{
A data.frame with 5 columns
}
\description{
This generic function accesses the eventData slot stored in an object derived 
from the Fast5Summary class.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   eventData( s.typhi.rep2 )
}
}

