% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-subsetting.R
\name{fastqTemplate}
\alias{fastqTemplate}
\title{Extract template reads}
\usage{
fastqTemplate(x)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{Fast5Summary}}}
}
\value{
A ShortReadQ object
}
\description{
This generic function accesses the fastq slot stored in an object derived 
from the Fast5Summary class, and returns only the template reads.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   fastqTemplate( s.typhi.rep2 )
}
}
