% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IONiseR.R
\docType{package}
\name{IONiseR}
\alias{IONiseR}
\alias{IONiseR-package}
\title{IONiseR: A package for assessing quality of MinION data}
\description{
IONiseR provides tools for the quality assessment of Oxford Nanopore MinION 
data. It extracts summary statistics from a set of fast5 files and can be 
used either before or after base calling.  In addition to standard summaries
of the read-types produced, it provides a number of plots for visualising 
metrics relative to experiment run time or spatially over the surface of a 
flowcell.
}

