% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/layoutPlots.R
\name{muxHeatmap}
\alias{muxHeatmap}
\title{Create layout plot of flowcell}
\usage{
muxHeatmap(data, zValue)
}
\arguments{
\item{data}{A data.frame or data.table.  Should have at least two columns,
one of which has the name 'channel'.}

\item{zValue}{Character string specifying the name of the column to be used
for the colour scaling.}
}
\value{
Returns an object of \code{gg} representing the plot.
}
\description{
Creates a plot representing the layout of a MinION flow cell.  Each circle
represents an individual channel with the intensity relecting a specified
sequencing metric.  This function is a more generalised version of
\code{\link{layoutPlot}}, allowing the user to map any value the like on
the channel layout.
}

