% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_summaryStats.R
\name{plotActiveChannels}
\alias{plotActiveChannels}
\title{Plot the number of active channels for each minute of run time}
\usage{
plotActiveChannels(summaryData)
}
\arguments{
\item{summaryData}{Object of class \linkS4class{Fast5Summary}.}
}
\value{
Returns an object of class \code{gg} representing the plot.
}
\description{
Plot the number of active channels for each minute of run time
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   plotActiveChannels( s.typhi.rep2 )
}
}
