% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_summaryStats.R
\name{plotCurrentByTime}
\alias{plotCurrentByTime}
\title{View changes in signal against run time.}
\usage{
plotCurrentByTime(summaryData)
}
\arguments{
\item{summaryData}{Object of class \linkS4class{Fast5Summary}.}
}
\value{
Returns an object of class \code{gg} representing the plot.
}
\description{
Plots the median recorded current for each fast5 file against the time at 
which the recording began.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   plotCurrentByTime( s.typhi.rep2 )
}
}

