% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_summaryStats.R
\name{plotEventRate}
\alias{plotEventRate}
\title{Plot the mean rate at which events occur}
\usage{
plotEventRate(summaryData)
}
\arguments{
\item{summaryData}{Object of class \linkS4class{Fast5Summary}.}
}
\value{
Returns an object of class \code{gg} representing the plot.
}
\description{
For each read, the ratio between the number of events comprising the read and the time spent in the pore is calculated.  This is then plotted against the time the read entered the pore, allow us to assess whether the rate at which events occur changes during the experiment run time.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   plotEventRate( s.typhi.rep2 )
}
}

