% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_summaryStats.R
\name{plotReadAccumulation}
\alias{plotReadAccumulation}
\title{Plot the accumulation of reads over the duration of the experiment.}
\usage{
plotReadAccumulation(summaryData)
}
\arguments{
\item{summaryData}{Object of class \linkS4class{Fast5Summary}.}
}
\value{
Returns an object of class \code{gg} representing the plot.
}
\description{
Plot the accumulation of reads over the duration of the experiment.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   plotReadAccumulation( s.typhi.rep2 )
}
}

