% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryPlots.R
\name{plotReadTypeProduction}
\alias{plotReadTypeProduction}
\title{View changes in signal against run time.}
\usage{
plotReadTypeProduction(summaryData, groupedMinutes = 10)
}
\arguments{
\item{summaryData}{Object of class \linkS4class{Fast5Summary}.}

\item{groupedMinutes}{Integer specifying how many minutes of runtime should be grouped together.}
}
\value{
Returns an object of class \code{gg} representing the plot.
}
\description{
Plots the median recorded current for each fast5 file against the time at
which the recording began.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   plotReadTypeProduction( s.typhi.rep2 )
}
}

