% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Methods-accessors.R
\name{rawData}
\alias{rawData}
\title{Extract rawData slot}
\usage{
rawData(x)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{Fast5Summary}}}
}
\value{
A data.table with 5 columns
}
\description{
This generic function accesses the rawData slot stored in an object derived
from the Fast5Summary class.
}
\examples{
if( require(minionSummaryData) ) {
   data(s.typhi.rep2, package = 'minionSummaryData')
   rawData( s.typhi.rep2 )
}
}

